pyproject.toml
[project]
name = "booker"
version = "0.1.0"
description = ""
authors = [
    {name = "onikageyoshi",email = "davidonyekachi29@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.9,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-rest-passwordreset (>=1.5.0,<2.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.42.1,<2.0.0)",
    "django-channels (>=0.7.0,<0.8.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "stripe (>=14.0.1,<15.0.0)",
    "razorpay (>=2.0.0,<3.0.0)",
    "paystackapi (>=2.1.3,<3.0.0)",
    "django-resized (>=1.0.3,<2.0.0)",
    "phonenumbers (>=9.0.19,<10.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "django-cloudinary-storage (>=0.3.0,<0.4.0)",
    "django-redis (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
